Use sleep-duration for Asleep; generic for InBed
OMH sleep-duration schema does not capture the state 'InBed', however we do not simply want to throw that data away when a HealthKit category sample has that value. Instead, HealthKit sleep analysis data that has a value of 'InBed' will be serialized using the generic hk-category-sample schema. HealthKit sleep analysis data that has a value of 'Asleep' will continue to be serialized using the OMH sleep-duration schema to increase data portability and interoperability.