Allow schema namespace to vary by serializer
Schemas can be from different namespaces so hardcoding a single namespace value did not work and would not support extensibility in the future. Each individual data point serializer now defines the name space that it belongs to and that value gets used in serializing the header for the data point.