Add support for all HealthKit types
We previously used a dictionary to map each HealthKit type identifier to the appropriate serializer. This was unnecessary and would not handle new types when added by Apple. We now maintain a dictionary that maps HealthKit types to OMH curated schemas when one exists that is a good match for the type and then default to the appropriate generic schema serializer when an OMH curated one does not exist.