Add dict with supported types and mapper classes
Previously we were using the prefix of the HK type (e.g., "HKQuantityType") to determine the appropriate mapper and then special casing for OMH supported types. However, it was problematic that there was no way of actually getting the list of types that are actually supported and their mappers. This is not only useful for ensuring we are clear about what we support, especially when apple updates the types, but is also useful for applications to be able to use that list when asking for user permissions or for their own purposes. The dictionary, contained within the OMHHealthKitConstantsMapper class, maps from HK types to mappers, either generic or type-specific. In addition, the OMHSerializer has a method that returns the list of keys which is, in essence, the list of supported types.