Serialize metadata date values as strings
The NSJSONSerialization class cannot serialize dates, so they need to be converted into strings before serializing. This would cause breaks in the serialization of metadata with dates. As a note, NSDate objects are the only types allowed as metadata that are not directly serializable by the class, so this should be the only special case needed.