Add back method to RFC3339 NSDate extension
A previous change removed the parameterless 'fromRFC3339String' method, which may be used by others. This change returns that and uses a UTC time zone in cases where one is not specified. Currently there is a rounding error with NSDate so two tests are failing, but the code is working correctly.